agent/requirements.txt
# Core LiveKit Agents Framework
livekit-agents>=0.6.0
livekit-plugins-openai>=0.6.0
livekit-plugins-silero>=0.6.0
livekit-plugins-cartesia>=0.1.0
livekit-plugins-assemblyai>=0.1.0
livekit-plugins-bvc>=0.1.0

# LiveKit SDK
livekit-server-sdk>=0.5.0
livekit-rtc>=0.3.0

# AI/ML Dependencies
openai>=1.0.0
assemblyai>=0.20.0

# Async Support
asyncio>=3.4.3
aiohttp>=3.9.0
httpx>=0.25.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
ipython>=8.0.0

# Monitoring
prometheus-client>=0.19.0

# Date/Time
python-dateutil>=2.8.0

# JSON handling
jsonschema>=4.0.0